requirements.txt
numpy
pandas
scikit-learn
opencv-python
scikit-image
Pillow
seaborn
matplotlib
tensorflow
sentence-transformers
transformers